Displays a matrixplot of a particular type of graphic from those included in the specimen for bivariate data in a html file.

A matrixplot is a grid of a particular type of graphic showing bivariate relationships between all pairs of variables of a certain(s) type(s) in a multivariate data set.

matrixplot(data, dataclass = NULL, diagram = NULL, dir = tempdir())

Arguments

data

Data.frame. Default dataset to use for plot. Unquoted. If not already a data.frame, it should be first coerced to by as.data.frame().

dataclass

Character vector. The types of data to be considered among the following:

'logical'

'ordered'

'factor'

'numeric'

'datetime'

'character'

diagram

Character. A specific graphic to be presented within the ones considered in the 2 input variables specimen available at https://sciencegraph.github.io/brinton/articles/specimen2.html.

dir

Directory in which the files are stored.

Value

Cause the side-effect of creating and displaying a temporary html file that includes a grid of graphics. The variables of a dataset are first grouped by the type of data, then, the variables matching the classes specified in the dataclass parameter, are represented in each row and/or column of the matrix.

See also

Examples

if (interactive()) {
matrixplot(iris, dataclass = c("numeric","numeric"),
diagram="bw contour plot with data points")
}