Presents a wideplot in a html output.

A wideplot is a grid of graphics were the graphics within each row corresponds to graphical representations of each one of the variables considered within a given dataset.

wideplot(data, dataclass = NULL, logical = NULL, ordered = NULL,
  factor = NULL, character = NULL, datetime = NULL, numeric = NULL,
  group = NULL, ncol = 7, label = "FALSE", dir = tempdir())

Arguments

data
Data.frame. Default dataset to use for plot. Unquoted. If not already a data.frame, it should be first coerced to by as.data.frame().

dataclass
Character vector. The types of data to be considered among the following:

'logical'

'ordered'

'factor'

'numeric'

'datetime'

'character'

logical
Character vector. Graphics for logical variables among the following:

'blank'

'line graph'

'tile plot'

'point graph'

'point-to-point graph'

'linerange graph'

'bar graph'

'bw bar graph'

'color bar graph'

'binned heatmap'

'bw binned heatmap'

'color binned heatmap'

ordered
Character vector. Graphics for ordered factor variables among the following:

'blank'

'line graph'

'tile plot'

'point graph'

'point-to-point graph'

'linerange graph'

'bar graph'

'bw bar graph'

'color bar graph'

'binned heatmap'

'bw binned heatmap'

'color binned heatmap'

factor
Character vector. Graphics for Character variables among the following:

'blank'

'line graph'

'freq. reordered line graph'

'alphab. reordered line graph'

'tile plot'

'freq. reordered tile plot'

'alphab. reordered tile plot'

'point graph'

'freq. reordered point graph'

'alphab. reordered point graph'

'binned heatmap'

'bw binned heatmap'

'color binned heatmap'

'freq. reordered binned heatmap'

'bw freq. reordered binned heatmap'

'color freq. reordered binned heatmap'

'alphab. reordered binned heatmap'

'bw alphab. reordered binned heatmap'

'color alphab. reordered binned heatmap'

'point-to-point graph'

'freq. reordered point-to-point graph'

'alphab. reordered point-to-point graph'

'linerange graph'

'freq. reordered linerange graph'

'alphab. reordered linerange graph'

'bar graph'

'bw bar graph'

'color bar graph'

'freq. reordered bar graph'

'bw freq. reordered bar graph'

'color freq. reordered bar graph'

'alphab. reordered bar graph'

'bw alphab. reordered bar graph'

'color alphab. reordered bar graph'

character
Character vector. Graphics for character variables among the following:

'blank'

'line graph'

'freq. reordered line graph'

'alphab. reordered line graph'

'tile plot'

'freq. reordered tile plot'

'alphab. reordered tile plot'

'point graph'

'freq. reordered point graph'

'alphab. reordered point graph'

'binned heatmap'

'bw binned heatmap'

'color binned heatmap'

'freq. reordered binned heatmap'

'bw freq. reordered binned heatmap'

'color freq. reordered binned heatmap'

'alphab. reordered binned heatmap'

'bw alphab. reordered binned heatmap'

'color alphab. reordered binned heatmap'

'point-to-point graph'

'freq. reordered point-to-point graph'

'alphab. reordered point-to-point graph'

'linerange graph'

'freq. reordered linerange graph'

'alphab. reordered linerange graph'

'bar graph'

'bw bar graph'

'color bar graph'

'freq. reordered bar graph'

'bw freq. reordered bar graph'

'color freq. reordered bar graph'

'alphab. reordered bar graph'

'bw alphab. reordered bar graph'

'color alphab. reordered bar graph'

datetime
Character vector. Graphics for datetime variables among the following:

'blank'

'line graph'

'stepped line graph'

'point graph'

'point-to-point graph'

'stepped point-to-point graph'

'binned heatmap'

'bw binned heatmap'

'color binned heatmap'

'bw heatmap'

'color heatmap'

numeric
Character vector. Graphics for numeric variables among the following:

'blank'

'area graph'

'stepped area graph'

'line graph'

'stepped line graph'

'stripe graph'

'bw stripe graph'

'color stripe graph'

'binned stripe graph'

'bw binned stripe graph'

'color binned stripe graph'

'point graph'

'bw point graph'

'color point graph'

'point graph with trend line'

'bw point graph with trend line'

'color point graph with trend line'

'binned heatmap'

'bw binned heatmap'

'color binned heatmap'

'bw heatmap'

'color heatmap'

'binned point graph'

'bw binned point graph'

'color binned point graph'

'point-to-point graph'

'stepped point-to-point graph'

'bar graph'

'bw bar graph'

'color bar graph'

'histogram'

'bw histogram'

'color histogram'

'density plot'

'filled density plot'

'violin plot'

'filled violin plot'

'box plot'

'3 uniaxial'

'normal qq plot'

group
Quoted character. Group of prestablished graphics which marks represent:

'sequence': the row number of the observations.

'scatter': graphics which marks represent individual observations.

'bin': observations into a series of intervals.

'model': a model built from the observations.

'symbol': different statistics through symbols.

'GOF': the goodness of fit of a statistical model.

'random': random graphics.

ncol
Numeric. Number of columns. An integer between 3 and 7. The fewer columns displayed, the larger the size of the resulting graphics, a feature that is especially useful if the scale labels dwarf the graphics area.

label
Logical. If `TRUE` the output includes labels that show the names of the graphics that are being displayed.

dir
Directory in which the files are stored.

Value

A html file that includes a grid of graphics. The variables of a dataset are first grouped by the type of data, then, each variable is graphically represented into a range of different graphics in one row of the matrix.

Examples

wideplot(sleep, dataclass = c("factor"), factor=c("point graph", "line graph", "tile plot"),
numeric = c("point graph", "line graph", "stepped line graph"))

#> 
#> 
#> processing file: wideplot.spin.Rmd

#>   |                                                                         |                                                                 |   0%  |                                                                         |.............                                                    |  20%
#>   ordinary text without R code
#> 
#>   |                                                                         |..........................                                       |  40%
#> label: preamble (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                         |.......................................                          |  60%
#> label: factor1 (with options) 
#> List of 2
#>  $ fig.width : num 13
#>  $ fig.height: num 0.8
#> 

#>   |                                                                         |....................................................             |  80%
#> label: factor2 (with options) 
#> List of 2
#>  $ fig.width : num 13
#>  $ fig.height: num 2.2
#> 

#>   |                                                                         |.................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: wideplot.knit.md

#> "C:/PROGRA~1/Pandoc/pandoc" +RTS -K512m -RTS wideplot.utf8.md --to html4 --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash+smart --output wideplot.html --email-obfuscation none --self-contained --standalone --section-divs --template "C:\Users\34601\Documents\R\win-library\3.4\rmarkdown\rmd\h\default.html" --no-highlight --variable highlightjs=1 --variable "theme:bootstrap" --include-in-header "C:\Users\34601\AppData\Local\Temp\RtmpkpIXB8\rmarkdown-str9dc1b484d2a.html" --mathjax --variable "mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML" 

#> 
#> Output created: wideplot.html