Presents a longplot in a html output.

A longplot is a range of suitable graphics that represent the relationship within the values of one, or a limited number, of variables in a dataset. Each graphic relates the values of all the selected variables and eventually the row number in which they appear.

longplot(data, vars, label = TRUE, dir = tempdir())

Arguments

data
Data.frame. Default dataset to use for plot. If not already a data.frame, it should be first coerced to by [as.data.frame()].

vars
Character. A specific variable within the dataset. Future work will allow to include a vector of variables.

label
Logical. If `TRUE` the output includes labels that show the names of the graphics that are being displayed.

dir
Directory in which the files are stored.

Value

A html file that includes a range of graphics suitable for this particular combination of variables.

Details

In order to present the range of graphics, the user must define a dataset and select one variable whitin it. Future work will include the ability to relate more than one variable.

Examples

# NOT RUN {
longplot(esoph, "tobgp")
# }