Displays a wideplot in a html file.

A wideplot is a grid of graphics where the graphics within each row corresponds to graphical representations of each one of the variables considered within a given dataset. The types of variables and the types of graphics are limited to those included in the specimen of graphics that require one input variable.

wideplot(
  data,
  dataclass = NULL,
  logical = NULL,
  ordered = NULL,
  factor = NULL,
  character = NULL,
  datetime = NULL,
  numeric = NULL,
  group = NULL,
  ncol = 7,
  label = "FALSE",
  dir = tempdir()
)

Arguments

data

Data.frame. Default dataset to use for plot. Unquoted. If not already a data.frame, it should be first coerced to by as.data.frame().

dataclass

Character vector. The types of data to be considered among the following:

'logical'

'ordered'

'factor'

'numeric'

'datetime'

'character'

logical

Character vector. Graphics for logical variables among the following:

'blank'

'line graph'

'tile plot'

'point graph'

'point-to-point graph'

'linerange graph'

'bar graph'

'bw bar graph'

'color bar graph'

'binned heatmap'

'bw binned heatmap'

'color binned heatmap'

ordered

Character vector. Graphics for ordered factor variables among the following:

'blank'

'line graph'

'tile plot'

'point graph'

'point-to-point graph'

'linerange graph'

'bar graph'

'bw bar graph'

'color bar graph'

'binned heatmap'

'bw binned heatmap'

'color binned heatmap'

factor

Character vector. Graphics for Character variables among the following:

'blank'

'line graph'

'freq. reordered line graph'

'alphab. reordered line graph'

'tile plot'

'freq. reordered tile plot'

'alphab. reordered tile plot'

'point graph'

'freq. reordered point graph'

'alphab. reordered point graph'

'binned heatmap'

'bw binned heatmap'

'color binned heatmap'

'freq. reordered binned heatmap'

'bw freq. reordered binned heatmap'

'color freq. reordered binned heatmap'

'alphab. reordered binned heatmap'

'bw alphab. reordered binned heatmap'

'color alphab. reordered binned heatmap'

'point-to-point graph'

'freq. reordered point-to-point graph'

'alphab. reordered point-to-point graph'

'linerange graph'

'freq. reordered linerange graph'

'alphab. reordered linerange graph'

'bar graph'

'bw bar graph'

'color bar graph'

'freq. reordered bar graph'

'bw freq. reordered bar graph'

'color freq. reordered bar graph'

'alphab. reordered bar graph'

'bw alphab. reordered bar graph'

'color alphab. reordered bar graph'

character

Character vector. Graphics for character variables among the following:

'blank'

'line graph'

'freq. reordered line graph'

'alphab. reordered line graph'

'tile plot'

'freq. reordered tile plot'

'alphab. reordered tile plot'

'point graph'

'freq. reordered point graph'

'alphab. reordered point graph'

'binned heatmap'

'bw binned heatmap'

'color binned heatmap'

'freq. reordered binned heatmap'

'bw freq. reordered binned heatmap'

'color freq. reordered binned heatmap'

'alphab. reordered binned heatmap'

'bw alphab. reordered binned heatmap'

'color alphab. reordered binned heatmap'

'point-to-point graph'

'freq. reordered point-to-point graph'

'alphab. reordered point-to-point graph'

'linerange graph'

'freq. reordered linerange graph'

'alphab. reordered linerange graph'

'bar graph'

'bw bar graph'

'color bar graph'

'freq. reordered bar graph'

'bw freq. reordered bar graph'

'color freq. reordered bar graph'

'alphab. reordered bar graph'

'bw alphab. reordered bar graph'

'color alphab. reordered bar graph'

datetime

Character vector. Graphics for datetime variables among the following:

'blank'

'line graph'

'stepped line graph'

'point graph'

'point-to-point graph'

'stepped point-to-point graph'

'binned heatmap'

'bw binned heatmap'

'color binned heatmap'

'bw heatmap'

'color heatmap'

numeric

Character vector. Graphics for numeric variables among the following:

'blank'

'area graph'

'stepped area graph'

'bw stepped area graph'

'color stepped area graph'

'seq. stripe graph'

'bw seq. stripe graph'

'color seq. stripe graph'

'line graph'

'stepped line graph'

'stripe graph'

'bw stripe graph'

'color stripe graph'

'binned stripe graph'

'bw binned stripe graph'

'color binned stripe graph'

'point graph'

'bw point graph'

'color point graph'

'point graph with trend line'

'bw point graph with trend line'

'color point graph with trend line'

'binned heatmap'

'bw binned heatmap'

'color binned heatmap'

'bw heatmap'

'color heatmap'

'binned point graph'

'bw binned point graph'

'color binned point graph'

'point-to-point graph'

'stepped point-to-point graph'

'bar graph'

'bw bar graph'

'color bar graph'

'histogram'

'bw histogram'

'color histogram'

'density plot'

'filled density plot'

'violin plot'

'filled violin plot'

'box plot'

'3 uniaxial'

'normal qq plot'

'ecdf plot'

'dotted ecdf plot'

'stepped ecdf plot'

group

Quoted character. Group of prestablished graphics which marks represent:

'sequence': the row number of the observations.

'scatter': graphics which marks represent individual observations.

'bin': observations into a series of intervals.

'model': a model built from the observations.

'symbol': different statistics through symbols.

'GOF': the goodness of fit of a statistical model.

'random': random graphics.

ncol

Numeric. Number of columns. An integer between 3 and 7. The fewer columns displayed, the larger the size of the resulting graphics, a feature that is especially useful if the scale labels dwarf the graphics area.

label

Logical. If `TRUE` the output includes labels that show the names of the graphics that are being displayed.

dir

Directory in which the files are stored.

Value

Cause the side-effect of creating and displaying a temporary html file that includes a grid of graphics. The variables of a dataset are first grouped by the type of data, then, each variable is graphically represented into a range of different graphics in one row of the matrix.

See also

Examples

if (interactive()) {
wideplot(sleep, dataclass = c("factor"),
factor=c("point graph", "line graph", "tile plot"),
numeric = c("point graph", "line graph", "stepped line graph"))
}